+++ ctor/tests/expand-darwin/ctor_anon.expanded.rs
@@ -2,6 +2,8 @@ use ctor::ctor;
 const _: () = {
     #[allow(dead_code)]
     unsafe fn foo() {
+        #[allow(unsafe_code)]
+        #[link_section = "__TEXT,__text_startup,regular,pure_instructions"]
         unsafe fn __ctor_private_inner() {
             {
                 ::std::io::_print(format_args!("foo\n"));
@@ -9,6 +11,7 @@ const _: () = {
         }
         const _: () = {
             #[allow(unsafe_code, unused_unsafe)]
+            #[link_section = "__TEXT,__text_startup,regular,pure_instructions"]
             extern "C" fn __ctor_private() {
                 { unsafe { __ctor_private_inner() } }
             }
@@ -25,6 +28,8 @@ const _: () = {
 const _: () = {
     #[allow(dead_code)]
     unsafe fn foo() {
+        #[allow(unsafe_code)]
+        #[link_section = "__TEXT,__text_startup,regular,pure_instructions"]
         unsafe fn __ctor_private_inner() {
             {
                 ::std::io::_print(format_args!("foo\n"));
@@ -32,6 +37,7 @@ const _: () = {
         }
         const _: () = {
             #[allow(unsafe_code, unused_unsafe)]
+            #[link_section = "__TEXT,__text_startup,regular,pure_instructions"]
             extern "C" fn __ctor_private() {
                 { unsafe { __ctor_private_inner() } }
             }