+++ ctor/tests/expand-darwin/ctor_naked.expanded.rs
@@ -1,6 +1,8 @@
 use ctor::ctor;
 #[allow(dead_code)]
 fn naked() {
+    #[allow(unsafe_code)]
+    #[link_section = "__TEXT,__text_startup,regular,pure_instructions"]
     fn __ctor_private_inner() {}
     const _: () = {
         #[allow(unsafe_code)]
@@ -8,6 +10,8 @@ fn naked() {
         #[used]
         static __CTOR_PRIVATE_REF: unsafe extern "C" fn() = {
             #[allow(unused_unsafe)]
+            #[allow(unsafe_code)]
+            #[link_section = "__TEXT,__text_startup,regular,pure_instructions"]
             extern "C" fn __ctor_private() {
                 { { __ctor_private_inner() } }
             }