+++ ctor/tests/expand-darwin/ctor_struct.expanded.rs
@@ -1,0 +1,5 @@
+struct Foo {}
+impl Foo {
+    #[ctor(unsafe)]
+    fn ctor() {}
+}