+++ ctor/tests/expand-darwin/ctor_unsafe.expanded.rs
@@ -1,6 +1,8 @@
 use ctor::ctor;
 #[allow(dead_code)]
 unsafe fn foo() {
+    #[allow(unsafe_code)]
+    #[link_section = "__TEXT,__text_startup,regular,pure_instructions"]
     unsafe fn __ctor_private_inner() {
         {
             ::std::io::_print(format_args!("foo\n"));
@@ -8,6 +10,7 @@ unsafe fn foo() {
     }
     const _: () = {
         #[allow(unsafe_code, unused_unsafe)]
+        #[link_section = "__TEXT,__text_startup,regular,pure_instructions"]
         extern "C" fn __ctor_private() {
             { unsafe { __ctor_private_inner() } }
         }