+++ ctor/tests/expand-linux/ctor_anon.expanded.rs
@@ -2,6 +2,8 @@ use ctor::ctor;
 const _: () = {
     #[allow(dead_code)]
     unsafe fn foo() {
+        #[allow(unsafe_code)]
+        #[link_section = ".text.startup"]
         unsafe fn __ctor_private_inner() {
             {
                 ::std::io::_print(format_args!("foo\n"));
@@ -13,6 +15,8 @@ const _: () = {
             #[used]
             static __CTOR_PRIVATE_REF: unsafe extern "C" fn() = {
                 #[allow(unused_unsafe)]
+                #[allow(unsafe_code)]
+                #[link_section = ".text.startup"]
                 extern "C" fn __ctor_private() {
                     { unsafe { __ctor_private_inner() } }
                 }
@@ -25,6 +29,8 @@ const _: () = {
 const _: () = {
     #[allow(dead_code)]
     unsafe fn foo() {
+        #[allow(unsafe_code)]
+        #[link_section = ".text.startup"]
         unsafe fn __ctor_private_inner() {
             {
                 ::std::io::_print(format_args!("foo\n"));
@@ -36,6 +42,8 @@ const _: () = {
             #[used]
             static __CTOR_PRIVATE_REF: unsafe extern "C" fn() = {
                 #[allow(unused_unsafe)]
+                #[allow(unsafe_code)]
+                #[link_section = ".text.startup"]
                 extern "C" fn __ctor_private() {
                     { unsafe { __ctor_private_inner() } }
                 }