+++ ctor/tests/expand-linux/ctor_attribute.expanded.rs
@@ -1,6 +1,8 @@
 use ctor::ctor;
 #[allow(dead_code)]
 fn foo() {
+    #[allow(unsafe_code)]
+    #[link_section = ".text.startup"]
     fn __ctor_private_inner() {
         {
             ::std::io::_print(format_args!("foo\n"));
@@ -12,6 +14,8 @@ fn foo() {
         #[used]
         static __CTOR_PRIVATE_REF: unsafe extern "C" fn() = {
             #[allow(unused_unsafe)]
+            #[allow(unsafe_code)]
+            #[link_section = ".text.startup"]
             extern "C" fn __ctor_private() {
                 { { __ctor_private_inner() } }
             }