+++ ctor/tests/expand-linux/ctor_doc.expanded.rs
@@ -3,6 +3,8 @@ use ctor::ctor;
 /// Doc 2
 #[allow(dead_code)]
 unsafe fn foo() {
+    #[allow(unsafe_code)]
+    #[link_section = ".text.startup"]
     unsafe fn __ctor_private_inner() {
         {
             ::std::io::_print(format_args!("foo\n"));
@@ -14,6 +16,8 @@ unsafe fn foo() {
         #[used]
         static __CTOR_PRIVATE_REF: unsafe extern "C" fn() = {
             #[allow(unused_unsafe)]
+            #[allow(unsafe_code)]
+            #[link_section = ".text.startup"]
             extern "C" fn __ctor_private() {
                 { unsafe { __ctor_private_inner() } }
             }