+++ ctor/tests/expand/ctor_multi_attr.rs
@@ -1,6 +1,6 @@
 use ctor::ctor;
 
-#[ctor(used(linker), link_section = ".ctors")]
+#[ctor(used(linker), link_section = ".ctors", body(link_section = ".text.startup"))]
 unsafe fn foo() {
     println!("foo");
 }