+++ ctor/tests/expand/ctor_struct.expanded.rs
@@ -1,7 +1,0 @@
-struct Foo {}
-impl Foo {
-    #[ctor(unsafe, link_section = ".ctors")]
-    fn ctor() {
-        (/*ERROR*/);
-    }
-}