+++ CHANGELOG.md
@@ -8,6 +8,10 @@ All notable changes to this project will be documented in this file.
 
 - `#[ctor(priority = naked)]` is now `#[ctor(naked)]`.
 
+### Added
+
+- Re-added link section option for body of `#[ctor]` items (supported for Linux/Android/FreeBSD/Apple).
+
 ## ctor [0.12.0] - 2026-04-30
 
 ### Added